updated spaaces for pointer uart calls
