+++ lib/uart/src/uart_ptr.c
@@ -1,12 +1,15 @@
+#pragma once
 #include "uart_ptr.h"
 
-void echo_uppercase_ptr(const struct device *dev)
+void echo_uppercase_ptr(const struct device *dev,
+                        int (*uart_ptr_in)(const struct device *, unsigned char *),
+                        void (*uart_ptr_out)(const struct device *, unsigned char *))
 {
     char byte, up;
 
     do
     {
-        if( (*uart_ptr_in)(dev, &byte) !=0 )
+        if( uart_ptr_in(dev, &byte) !=0 )
             continue;
 
         if (byte <= 'z' && byte >= 'a')
@@ -15,7 +18,7 @@ void echo_uppercase_ptr(const struct device *dev)
         else
             up = byte;
 
-        (*uart_ptr_out)(dev,up)
+        (*uart_ptr_out)(dev,up);
     }
     while(byte != '\n');
