+++ platformio.ini
@@ -15,7 +15,7 @@ framework = zephyr
 platform_packages =
     platformio/tool-renode
 test_testing_command =
-    ${platformio.packages_dir}/tool-renode/renode-test
+    ${platformio.packages_dir}/tool-renode/bin/renode-test
     test/uart_echo.robot
 test_ignore = mocked/*
 
@@ -29,7 +29,7 @@ build_flags =
 platform_packages =
     platformio/tool-renode
 test_testing_command =
-    ${platformio.packages_dir}/tool-renode/renode
+    ${platformio.packages_dir}/tool-renode/bin/renode
     --disable-xwt
     -e mach create "stm32f072b"
     -e machine LoadPlatformDescription @platforms/boards/stm32f072b_discovery.repl
