+++ test/test_ptr/include/test_uart_ptr.h
@@ -1,0 +1,3 @@
+
+int test_uart_out();
+int test_uart_in();
