+++ zephyr/prj.conf
@@ -1,0 +1,2 @@
+CONFIG_SERIAL=y
+CONFIG_NEWLIB_LIBC=y
